chore: 디렉토리, 스토리북 및 스타일 로직 분리
