+++ src/components/common/CheckBox.tsx
@@ -1,5 +1,5 @@
 import { useId } from 'react';
-import styled from 'styled-components';
+import * as S from './styles/CheckBox.styles';
 
 interface Props {
   checked: boolean;
@@ -10,37 +10,11 @@ export default function CheckBox({ checked, onClickCheckbox }: Props) {
   const id = useId();
 
   return (
-    <Wrapper>
+    <S.Wrapper>
       <input id={id} type="checkbox" checked={checked} />
-      <CheckboxLabel htmlFor={id} onClick={onClickCheckbox}>
+      <S.CheckboxLabel htmlFor={id} onClick={onClickCheckbox}>
         {checked && <img src="./whiteCheck.svg" />}
-      </CheckboxLabel>
-    </Wrapper>
+      </S.CheckboxLabel>
+    </S.Wrapper>
   );
 }
-
-const Wrapper = styled.div`
-  input {
-    display: none;
-  }
-`;
-
-const CheckboxLabel = styled.label`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 28px;
-  height: 28px;
-  border: 1.5px solid #22a6a2;
-  border-radius: 2px;
-
-  &:hover {
-    cursor: pointer;
-  }
-
-  input:checked ~ & {
-    border-color: #3288ff;
-    background: #333333;
-  }
-`;
