+++ src/components/common/QuantityInput.tsx
@@ -1,10 +1,10 @@
 import { useEffect } from 'react';
 import { useRecoilValue } from 'recoil';
-import styled from 'styled-components';
+import * as S from './styles/QuantityInput.styles';
 
-import { cartItemState } from '../../recoil/state';
-import useQuantityInput from '../../hooks/useQuantityInput';
-import { isNaturalNumberString } from '../../utils/validator';
+import { cartItemState } from '../../atom/state';
+import useQuantityInput from '../hooks/useQuantityInput';
+import { checkNaturalStringTypeNumber } from '../../utils/checkNaturalStringTypeNumber';
 
 interface Props {
   cartItemId: number;
@@ -24,7 +24,7 @@ export default function QuantityInput({ cartItemId, min = 0, max, style }: Props
   };
 
   const onChangeInput = ({ target: { value } }: React.ChangeEvent<HTMLInputElement>) => {
-    if (isNaturalNumberString(value)) {
+    if (checkNaturalStringTypeNumber(value)) {
       setInputWithRequest(getValidRange(Number(value)));
     } else if (value === '') {
       setInput(value);
@@ -48,69 +48,16 @@ export default function QuantityInput({ cartItemId, min = 0, max, style }: Props
   }, []);
 
   return (
-    <Wrapper style={style}>
-      <Input type="text" value={input} onChange={onChangeInput} onBlur={onBlurInput} />
-      <CounterBox>
-        <Counter onClick={quantityIncrease} disabled={Number(input) === max}>
+    <S.Wrapper style={style}>
+      <S.Input type="text" value={input} onChange={onChangeInput} onBlur={onBlurInput} />
+      <S.CounterBox>
+        <S.Counter onClick={quantityIncrease} disabled={Number(input) === max}>
           <img src="./arrowUp.svg" />
-        </Counter>
-        <Counter onClick={quantityDecrease} disabled={Number(input) === min}>
+        </S.Counter>
+        <S.Counter onClick={quantityDecrease} disabled={Number(input) === min}>
           <img src="./arrowDown.svg" />
-        </Counter>
-      </CounterBox>
-    </Wrapper>
+        </S.Counter>
+      </S.CounterBox>
+    </S.Wrapper>
   );
 }
-
-const Wrapper = styled.div`
-  display: flex;
-
-  width: 72px;
-  height: 36px;
-
-  text-align: center;
-  font-size: 16px;
-`;
-
-const Input = styled.input`
-  width: 64%;
-  height: 100%;
-  border: 1px solid #dddddd;
-
-  text-align: center;
-  font-size: inherit;
-  color: #333333;
-`;
-
-const CounterBox = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  width: 36%;
-  height: 100%;
-`;
-
-const Counter = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 50%;
-  border: 1px solid #dddddd;
-  background: transparent;
-
-  & > img {
-    width: 48%;
-    height: 32%;
-  }
-
-  &:disabled {
-    background: rgba(0, 0, 0, 0.1);
-    cursor: default;
-  }
-
-  &:disabled > img {
-    visibility: hidden;
-  }
-`;
