+++ src/components/productList/Product.tsx
@@ -1,14 +1,11 @@
 import type { ProductType } from '../../types';
-
 import { useState } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import styled from 'styled-components';
-
+import * as S from './styles/Product.styles';
 import QuantityInput from '../common/QuantityInput';
-
 import * as api from '../../api';
-import useToast from '../../hooks/useToast';
-import { cartState, serverNameState } from '../../recoil/state';
+import useToast from '../hooks/useToast';
+import { cartState, serverNameState } from '../../atom/state';
 import { API_ERROR_MESSAGE, API_SUCCESS_MESSAGE, MAX_QUANTITY } from '../../constants';
 
 interface Props extends ProductType {}
@@ -48,96 +45,24 @@ export default function Product({ id, name, price, imageUrl }: Props) {
 
   return (
     <>
-      <Wrapper>
-        <Image src={imageUrl} onError={setAltSrc} />
-        <InfoBox>
-          <LabelBox>
-            <Name>{name}</Name>
-            <Price>{price.toLocaleString()} 원</Price>
-          </LabelBox>
-          <ControlBox>
+      <S.Wrapper>
+        <S.Image src={imageUrl} onError={setAltSrc} />
+        <S.InfoBox>
+          <S.LabelBox>
+            <S.Name>{name}</S.Name>
+            <S.Price>{price.toLocaleString()} 원</S.Price>
+          </S.LabelBox>
+          <S.ControlBox>
             {cartItem ? (
               <QuantityInput cartItemId={cartItem.id} min={0} max={MAX_QUANTITY} />
             ) : (
-              <CartItemAddButton onClick={addCartItem} disabled={addLoading}>
+              <S.CartItemAddButton onClick={addCartItem} disabled={addLoading}>
                 <img src="./cart.svg" />
-              </CartItemAddButton>
+              </S.CartItemAddButton>
             )}
-          </ControlBox>
-        </InfoBox>
-      </Wrapper>
+          </S.ControlBox>
+        </S.InfoBox>
+      </S.Wrapper>
     </>
   );
 }
-
-const Wrapper = styled.div`
-  width: 282px;
-  height: 362px;
-
-  color: #333333;
-`;
-
-const Image = styled.img`
-  width: 100%;
-  height: 282px;
-`;
-
-const InfoBox = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  width: 282px;
-  padding-top: 16px;
-  padding-left: 8px;
-`;
-
-const LabelBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`;
-
-const Name = styled.p`
-  margin-top: 4px;
-
-  font-size: 16px;
-  font-weight: 400;
-`;
-
-const Price = styled.p`
-  margin-top: 10px;
-
-  vertical-align: center;
-  font-size: 20px;
-  font-weight: 600;
-`;
-
-const CartItemAddButton = styled.button`
-  width: 32px;
-  height: 30px;
-  margin-right: 10px;
-
-  background: transparent;
-
-  transition: transform 0.2s;
-
-  & > img {
-    width: 100%;
-    height: 100%;
-  }
-
-  &:hover {
-    transform: rotate(12deg);
-  }
-
-  &:disabled {
-    cursor: wait;
-  }
-  &:disabled > img {
-    visibility: hidden;
-  }
-`;
-
-const ControlBox = styled.div`
-  width: auto;
-`;
