+++ src/components/productList/ProductList.tsx
@@ -1,15 +1,12 @@
 import type { ProductType } from '../../types';
-
 import { useEffect, useState } from 'react';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
-import styled from 'styled-components';
-
+import * as S from './styles/ProductList.styles';
 import Product from './Product';
 import SkeletonProduct from './SkeletonProduct';
-
 import * as api from '../../api';
-import useToast from '../../hooks/useToast';
-import { cartState, serverNameState } from '../../recoil/state';
+import useToast from '../hooks/useToast';
+import { cartState, serverNameState } from '../../atom/state';
 import { API_ERROR_MESSAGE, SKELETONS_LENGTH } from '../../constants';
 
 export default function ProductList() {
@@ -35,29 +32,10 @@ export default function ProductList() {
   }, [serverName]);
 
   return (
-    <Wrapper>
+    <S.Wrapper>
       {products === null
         ? Array.from({ length: SKELETONS_LENGTH }).map(() => <SkeletonProduct />)
         : products.map((product) => <Product key={product.id} {...product} />)}
-    </Wrapper>
+    </S.Wrapper>
   );
 }
-
-const Wrapper = styled.div`
-  display: grid;
-  grid-template-columns: repeat(4, 282px);
-  grid-column-gap: 48px;
-  grid-row-gap: 64px;
-
-  @media (max-width: 1272px) {
-    grid-template-columns: repeat(3, 282px);
-  }
-
-  @media (max-width: 942px) {
-    grid-template-columns: repeat(2, 282px);
-  }
-
-  @media (max-width: 612px) {
-    grid-template-columns: repeat(1, 282px);
-  }
-`;
