+++ src/page/styles/ErrorPage.styles.ts
@@ -1,19 +1,7 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-export default function ErrorPage() {
-  return (
-    <Wrapper>
-      <Image src="./404.svg" />
-      <Title>페이지를 찾지 못했어요.</Title>
-      <Message>주소에 오타가 있거나, </Message>
-      <Message>변경, 삭제되었을 수도 있어요!</Message>
-      <StyledLink to="/">홈으로 가기</StyledLink>
-    </Wrapper>
-  );
-}
-
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -25,23 +13,23 @@ const Wrapper = styled.div`
   font-size: 16px;
 `;
 
-const Image = styled.img`
+export const Image = styled.img`
   margin-bottom: 16px;
 `;
 
-const Title = styled.h1`
+export const Title = styled.h1`
   margin: 32px 0;
 
   font-size: 28px;
 `;
 
-const Message = styled.p`
+export const Message = styled.p`
   margin: 8px 0;
 
   font-size: 16px;
 `;
 
-const StyledLink = styled(Link)`
+export const StyledLink = styled(Link)`
   display: flex;
   justify-content: center;
   align-items: center;
