+++ src/stories/Product.stories.tsx
@@ -1,6 +1,5 @@
 import type { StoryFn } from '@storybook/react';
-
-import Product from './Product';
+import Product from '../components/productList/Product';
 
 export default {
   title: 'Product',
