+++ src/utils/checkNaturalStringTypeNumber.ts
@@ -1,3 +1,3 @@
-export const isNaturalNumberString = (string: string) => {
+export const checkNaturalStringTypeNumber = (string: string) => {
   return string !== '' && string.split('').every((char) => '0123456789'.includes(char));
 };
